+++ pyonepassword/_op_cli_config.py
@@ -1,7 +1,8 @@
 import json
+import logging
 import os
-import pathlib
 from json.decoder import JSONDecodeError
+from pathlib import Path
 from typing import List, Optional
 
 from .py_op_exceptions import OPConfigNotFoundException
@@ -34,15 +35,16 @@ class OPCLIAccountConfig(dict):
 
 
 class OPCLIConfig(dict):
-    OP_CONFIG_PATHS = [
-        pathlib.Path(".config", "op", "config"),
-        pathlib.Path(".op", "config")
-    ]
 
-    def __init__(self, configpath=None):
+    def __init__(self, config_dir: Optional[str | Path] = None, logger: Optional[logging.Logger] = None):
         super().__init__()
-        if configpath is None:
-            configpath = self._get_config_path()
+        if not logger:
+            logger = logging.getLogger(self.__class__.__name__)
+            logger.setLevel(logging.INFO)
+        self.logger = logger
+
+        configpath = self._get_config_path(config_dir=config_dir)
+        self.logger.debug(f"configpath: {configpath}")
         self.configpath = configpath
         if configpath is None:
             raise OPConfigNotFoundException("No op configuration found")
@@ -71,22 +73,138 @@ class OPCLIConfig(dict):
             account_map[account.shorthand] = account
         self.account_map = account_map
 
-    def _get_config_path(self) -> pathlib.Path:
-        configpath: pathlib.Path = None
-        config_home = None
-        try:
-            config_home = pathlib.Path(os.environ['XDG_CONFIG_HOME'])
+    def _get_config_path(self, config_dir=None) -> Path | None:
+        """
+        Determine the path to the 1Password CLI configuration file.
 
-        except KeyError:
-            config_home = pathlib.Path.home()
+        This method follows the 1Password CLI configuration directory rules
+        to determine the order in which configuration files should be checked.
+        The rules are applied in order:
+        1. A directory specified with config_dir
+        2. A directory set with the OP_CONFIG_DIR environment variable
+        3. ~/.op
+        4. ${XDG_CONFIG_HOME}/.op
+        5. ~/.config/op
+        6. ${XDG_CONFIG_HOME}/op
 
-        for subpath in self.OP_CONFIG_PATHS:
-            _configpath = pathlib.Path(config_home, subpath)
-            if os.path.exists(_configpath):
+        NOTE: If a custom configuration directory is provided or OP_CONFIG_DIR is set,
+        their use is mandatory and there is no fallback if the config file isn't present.
+
+        Args:
+            config_dir: Optional custom configuration directory path.
+                        If provided, this will be used as the base directory
+                        for locating the config file.
+
+        Returns:
+            Path: The path to the configuration file if found, otherwise None.
+        """
+        configpath: Path | None = None
+        path_options = self._config_path_triage_list(config_dir)
+        for _configpath in path_options:
+            self.logger.debug(f"Looking for config at {_configpath}")
+            if _configpath.exists():
                 configpath = _configpath
                 break
 
         return configpath
+
+    def _get_custom_config_dir(self, custom_config_dir: str | Path) -> Path | None:
+        """
+        Determine the custom configuration directory path if one has been specified.
+
+        If the user has specified a custom configuration directory, we must use that,
+        and there's no fallback if it doesn't exist.
+
+        NOTE: This is a configuration directory. A config file named "config" will be
+        looked for inside.
+
+        This can take the form of (in order):
+        - explicitly provided
+        - set in OP_CONFIG_DIR env variable.
+
+        Args:
+            custom_config_dir: The custom configuration directory path, if provided.
+
+        Returns:
+            Path: The custom configuration directory path, or None if not specified.
+        """
+        config_dir_return: Path | None = None
+        if not custom_config_dir:
+            op_conf_dir = os.environ.get("OP_CONFIG_DIR", None)
+            if op_conf_dir:
+                config_dir_return = Path(op_conf_dir)
+                self.logger.debug(f"OP_CONFIG_DIR set to: {custom_config_dir}")
+        else:
+            config_dir_return = Path(custom_config_dir)
+            self.logger.debug(
+                f"Custom config dir specified: {custom_config_dir}")
+        return config_dir_return
+
+    def _config_path_triage_list(self, custom_config_dir) -> list[Path]:
+        """
+        Generate a list of potential configuration file paths to check.
+
+
+        Args:
+            custom_config_dir: The custom configuration directory path, if provided.
+
+        Returns:
+            list[Path]: A list of potential configuration file paths to check,
+                       in order of preference.
+        """
+        xdg_home = os.environ.get('XDG_CONFIG_HOME', None)
+        if xdg_home:
+            self.logger.debug(f"XDG_CONFIG_HOME set to {xdg_home}")
+        custom_config_dir = self._get_custom_config_dir(custom_config_dir)
+        configpath = None
+        path_options: list[Path] = []
+        # Rules from https://developer.1password.com/docs/cli/config-directories/
+        if custom_config_dir:
+            # rule 1: A directory specified with --config
+            # rule 2: A directory set with the OP_CONFIG_DIR environment variable.
+            configpath = Path(custom_config_dir, "config")
+            self.logger.debug(
+                f"Rule 1 & 2: Adding custom_config_dir to path triage list: {configpath}")
+
+            if not configpath.exists():
+                # we were explicitly told to use this path, so there's no fallback
+                # if it doesn't exist
+                raise OPConfigNotFoundException(
+                    f"Caller-provided config not found at {configpath}")
+            else:
+                path_options.append(configpath)
+
+        else:
+            # rule 1 or 2 weren't met, so we evaluate the rest
+            # rule 3: ~/.op (following go-homedir  to determine the home directory)
+            # We use pathlib.Path.expanduser(). Hopefully this matches go-homedir semantics
+            self.logger.debug("Rule 3: adding ~/.op to path triage list")
+            config_dir = Path("~/.op").expanduser()
+            config_path = Path(config_dir, "config")
+            path_options.append(config_path)
+
+            # rule 4: ${XDG_CONFIG_HOME}/.op
+            if xdg_home:
+                self.logger.debug(
+                    "Rule 4: Adding ${XDG_CONFIG_HOME}/.op to path triage list")
+                config_path = Path(xdg_home, ".op", "config")
+                path_options.append(config_path)
+
+            # rule 5: ~/.config/op (following go-homedir to determine the home directory)
+            self.logger.debug(
+                "Rule 5: Adding ~/.config/op to path triage list")
+            config_dir = Path("~/.config", "op").expanduser()
+            config_path = Path(config_dir, "config")
+            path_options.append(config_path)
+
+            # rule 6: ${XDG_CONFIG_HOME}/op
+            if xdg_home:
+                self.logger.debug(
+                    "Rule 6: Adding ${XDG_CONFIG_HOME}/op to path triage list")
+                config_path = Path(xdg_home, "op", "config")
+                path_options.append(config_path)
+
+        return path_options
 
     def _initialize_account_objects(self) -> List[OPCLIAccountConfig]:
         account_list = self.accounts